+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -876,7 +876,6 @@ class FrameworkExtension extends Extension
         if (null !== $jsonManifestPath) {
             $def = new ChildDefinition('assets.json_manifest_version_strategy');
             $def->replaceArgument(0, $jsonManifestPath);
-            $def->setPublic(false);
             $container->setDefinition('assets._version_'.$name, $def);
 
             return new Reference('assets._version_'.$name);