+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 -----
 
  * Added a new new version strategy option called json_manifest_path
-   that allows you to use the JsonManifestVersionStrategy.
+   that allows you to use the `JsonManifestVersionStrategy`.
  * Added support for the `controller.service_arguments` tag, for injecting services into controllers' actions
  * Deprecated `cache:clear` with warmup (always call it with `--no-warmup`)
  * Deprecated the "framework.trusted_proxies" configuration option and the corresponding "kernel.trusted_proxies" parameter