+++ src/Symfony/Component/Asset/CHANGELOG.md
@@ -3,7 +3,7 @@ CHANGELOG
 
 3.3.0
 -----
- * Added JsonManifestVersionStrategy as a way to read final,
+ * Added `JsonManifestVersionStrategy` as a way to read final,
    versioned paths from a JSON manifest file.
 
 2.7.0